Introduce buildUrl to safely build URLs to the incident.io dashboard
It seems like `URLSearchParams.toString` includes a `?` on some
platforms and does not include it on others, although I can't find a
source that agrees with this.

This attempts to work around that by building a `URL` object and then
`.toString`-ing that, which hopefully means that the platform matches
the behaviour of `URL.searchParams` and setting `URL.search`.