Bump backstage/cli to 0.34.3 per dependabot
Bumps backstage/cli to 0.34.3 per dependabot. Bumping this and running `yarn install` causes `yarn lint` to fail. To make it work it required following the steps in the jsx-transform-migration [page](https://backstage.io/docs/tutorials/jsx-transform-migration/). Importantly, the compiler option of "react-jsx" requires React 17 or higher, so we updated the React dependencies to account for this in a previous commit.